+++ conan/run_in_docker.py
@@ -1,7 +1,7 @@
 import os
 from conan.log import logger
 
-from conan.tst_package_runner import DockerTestPackageRunner
+from conan.test_package_runner import DockerTestPackageRunner
 
 
 def run():
